refactor: move out setting up logging into sep package
